refactor: split expansion.c
