+++ src/expansion/expansion.c
@@ -1,121 +1,10 @@
 #include "minishell.h"
 
-size_t	get_word_len(char *str, bool in_squote, bool in_dquote)
-{
-	char	*head;
-
-	head = str;
-	if ((in_squote && *str == '\'') || (!in_squote && *str == '\"'))
-		return (1);
-	if (in_squote)
-		return (ft_strchr(str, '\'') - str);
-	if (*str == '$')
-	{
-		str++;
-		if (*str == '?')
-			return (2);
-		while (*str && (ft_isalnum(*str) || *str == '_'))
-			str++;
-	}
-	else if (in_dquote)
-		while (*str && *str != '$' && *str != '\"')
-			str++;
-	else if(ft_isspace(*str))
-		while (*str && ft_isspace(*str))
-			str++;
-	else
-		while (*str && !ft_strchr("\'\"$ ", *str))
-			str++;
-	return (str - head);
-}
-
-t_list	*extract_word(char **str, bool in_squote, bool in_dquote, t_expansion_kind kind)
-{
-	t_list		*new;
-	t_expansion	*exp;
-
-	exp = ft_calloc(1, sizeof(t_expansion));
-	exp->str = ft_substr(*str, 0, get_word_len(*str, in_squote, in_dquote));
-	exp->len = ft_strlen(exp->str);
-	*str += exp->len;
-	exp->in_squote = in_squote;
-	exp->in_dquote = in_dquote;
-	exp->kind = kind;
-	if (exp->str && exp->str[0] == '$' && (ft_isalnum(exp->str[1]) || exp->str[1] == '_' || exp->str[1] == '?') && !in_squote)
-		exp->kind = ENV;
-	if (exp->str && exp->str[0] == ' ' && !in_squote && !in_dquote)
-		exp->kind = NAKED_SPACE;
-	new = ft_lstnew(exp);
-	return (new);
-}
-
-t_list	*split_str(char *str, bool par_in_dquote)
-{
-	t_list	*head;
-	bool	in_dquote;
-
-	if (!str)
-		return (NULL);
-	head = NULL;
-	if (*str == '\0')
-		ft_lstadd_back(&head, extract_word(&str, false, par_in_dquote, STRING));
-	in_dquote = false;
-	while (*str)
-	{
-		if (!in_dquote && *str == '\'')
-		{
-			ft_lstadd_back(&head, extract_word(&str, true, par_in_dquote | in_dquote, SQUOTE));
-			ft_lstadd_back(&head, extract_word(&str, true, par_in_dquote | in_dquote, STRING));
-			ft_lstadd_back(&head, extract_word(&str, true, par_in_dquote | in_dquote, SQUOTE));
-		}
-		else if (*str == '\"')
-		{
-			in_dquote ^= 1;
-			ft_lstadd_back(&head, extract_word(&str, false, true, DQUOTE));
-		}
-		else
-			ft_lstadd_back(&head, extract_word(&str, false, par_in_dquote | in_dquote, STRING));
-	}
-	return (head);
-}
-
-t_list	*get_expansion_list(char *str, bool par_in_dquote, t_shell_var *shell_var)
-{
-	t_list	*head;
-	t_list	*itr;
-	t_list	*prev;
-	t_list	*next;
-	t_expansion	*exp;
-
-	prev = NULL;
-	head = split_str(str, par_in_dquote);
-	itr = head;
-	while (itr)
-	{
-		next = itr->next;
-		exp = (t_expansion *)(itr->content);
-		if (exp->kind == ENV)
-		{
-			exp->str = get_env_value_str(exp->str + 1, shell_var);
-			if (prev == NULL)
-				head = get_expansion_list(exp->str, exp->in_dquote, shell_var);
-			else
-				prev->next = get_expansion_list(exp->str, exp->in_dquote, shell_var);
-			prev = ft_lstlast(head);
-			prev->next = next;
-		}
-		else
-			prev = itr;
-		itr = next;
-	}
-	return (head);
-}
-
 t_list	*remove_quotes(t_list *src_list)
 {
-	t_list	*head;
-	t_list	*itr;
-	t_list	*next;
+	t_list		*head;
+	t_list		*itr;
+	t_list		*next;
 	t_expansion	*exp;
 
 	head = NULL;
@@ -126,7 +15,7 @@ t_list	*remove_quotes(t_list *src_list)
 		itr->next = NULL;
 		exp = (t_expansion *)(itr->content);
 		if (exp->kind == SQUOTE || exp->kind == DQUOTE)
-			; // exp_strlist_delone(itr);
+			ft_lstdelone(itr, delete_expansion);
 		else
 			ft_lstadd_back(&head, itr);
 		itr = next;
@@ -164,7 +53,8 @@ char	*token_str_join(t_list **src_list, char *buf)
 	itr = *src_list;
 	if (((t_expansion *)(itr->content))->kind == FILENAME_EXPANSION)
 	{
-		ft_strlcat(buf, ((t_expansion *)(itr->content))->str, ((t_expansion *)(itr->content))->len + 1);
+		exp = (t_expansion *)(itr->content);
+		ft_strlcat(buf, exp->str, exp->len + 1);
 		*src_list = (*src_list)->next;
 		return (buf);
 	}
@@ -181,7 +71,7 @@ char	*token_str_join(t_list **src_list, char *buf)
 }
 
 // naked spaceで分割し、文字列(t_exp_strlist)を連結してトークン化する
-t_list *convert_to_token(t_list *expansion_list)
+t_list	*convert_to_token(t_list *expansion_list)
 {
 	t_list	*head; // token_list
 	t_list	*itr;
@@ -216,6 +106,7 @@ t_list	*get_expanded_token(t_list *token_list, t_shell_var *shell_var)
 	t_list			*expansion_list;
 	t_list			*expanded_token_list;
 
+	// token_list->next = NULL;
 	// token->strをt_exp_strlist_typeで分割・分類する
 	// 環境変数の展開
 	expansion_list = get_expansion_list(((t_token *)(token_list->content))->str, false, shell_var); // t_expansionのリスト
@@ -223,47 +114,51 @@ t_list	*get_expanded_token(t_list *token_list, t_shell_var *shell_var)
 	expansion_list = get_filename_expansion(expansion_list);
 	expansion_list = remove_quotes(expansion_list); // t_expansionのリスト
 	expanded_token_list = convert_to_token(expansion_list); // t_tokenのリスト
-	// printf("226: |%s|\n", ((t_token *)(expanded_token_list->content))->str);
+	ft_lstdelone(token_list, delete_token);
+	ft_lstclear(&expansion_list, delete_expansion);
 	return (expanded_token_list);
 }
 
-void	handle_process(t_process *process, t_shell_var *shell_var)
+void	handle_process(t_list **token_list, t_shell_var *shell_var)
 {
-	t_list	*itr; //token_list_itr
+	t_list	head;
+	t_list	*itr;
 	t_list	*next;
 	t_list	*prev;
 	bool	is_delimiter;
 
-	itr = process->token_list;
-	prev = NULL;
+	head.next = *token_list;
+	itr = *token_list;
+	prev = &head;
 	is_delimiter = false;
 	while (itr)
 	{
 		next = itr->next;
+		itr->next = NULL;
 		if (!is_delimiter && ((t_token *)(itr->content))->kind == TK_STRING)
-			itr = get_expanded_token(itr, shell_var);
-		if (itr == NULL)
 		{
-			itr = next;
-			if (prev)
+			itr = get_expanded_token(itr, shell_var);
+			if (itr == NULL)
+			{
+				itr = next;
 				prev->next = itr;
-			else
-				process->token_list = itr;
-			continue ;
-		}
-		if (prev)
+				continue ;
+			}
 			prev->next = itr;
+			itr = ft_lstlast(itr);
+		}
 		else
-			process->token_list = itr;
-		while (itr->next != NULL && itr->next != next)
-			itr = itr->next;
-		is_delimiter = false;
-		if (((t_token *)(itr->content))->kind == TK_REDIRECT && ft_strcmp(((t_token *)(itr->content))->str, "<<") == 0)
-			is_delimiter = true;
+		{
+			is_delimiter = false;
+			if (((t_token *)(itr->content))->kind == TK_REDIRECT && \
+				ft_strcmp(((t_token *)(itr->content))->str, "<<") == 0)
+				is_delimiter = true;
+		}
 		itr->next = next;
 		prev = itr;
 		itr = itr->next;
 	}
+	*token_list = head.next;
 }
 
 void	expansion(t_expression *expression, t_shell_var *shell_var)
@@ -273,7 +168,7 @@ void	expansion(t_expression *expression, t_shell_var *shell_var)
 	itr = expression->process_list;
 	while(itr)
 	{
-		handle_process((t_process *)(itr->content), shell_var);
+		handle_process(&((t_process *)(itr->content))->token_list, shell_var);
 		g_exit_status = 0; // bashの挙動に合わせた
 		itr = itr->next;
 	}
